+++ app/setup/openSansObserver.js
